refactor: Move posthog client to its own file
