+++ services/mcp/src/lib/errors.ts
@@ -1,6 +1,6 @@
 import type { CallToolResult } from '@modelcontextprotocol/sdk/types.js'
 
-import { getPostHogClient } from '@/lib/posthog/analytics'
+import { getPostHogClient } from '@/lib/posthog'
 import { sanitizeHeaderValue } from '@/lib/utils'
 
 export enum ErrorCode {
