+++ services/mcp/src/lib/posthog/index.ts
@@ -1,0 +1,3 @@
+import { getPostHogClient } from './client'
+
+export { getPostHogClient }
