+++ services/mcp/tests/unit/mcp-cache.test.ts
@@ -6,11 +6,13 @@ import type { State } from '@/tools/types'
 
 const mockCaptureException = vi.fn()
 
-vi.mock('@/lib/posthog/analytics', () => ({
+vi.mock('@/lib/posthog', () => ({
     getPostHogClient: () => ({
         captureException: mockCaptureException,
         capture: vi.fn(),
     }),
+}))
+vi.mock('@/lib/posthog/analytics', () => ({
     AnalyticsEvent: {},
 }))
 
@@ -27,7 +29,7 @@ async function getOrFetchCached<D extends keyof State, T extends keyof State>(de
         fetcher: () => Promise<State[D]>
     }
 }): Promise<State[D] | undefined> {
-    const { getPostHogClient } = await import('@/lib/posthog/analytics')
+    const { getPostHogClient } = await import('@/lib/posthog')
     const { cache, waitUntil, cacheTtlMs, opts } = deps
 
     try {
